// src/app/core/types/reports/BundledEODStats.ts
export interface BundledEODStats {

    totalApplicationAtBranch: number
    totalApplicationDocChecked: number
    totalApplicationInProgress: number
    totalApplicationApproved: number
    totalApplicationRejected: number
    totalApplicationPending: number
    totalApplicationReturnToBranch: number
    totalApplicationFinished: number
    totalApplicationCDDReviewCheck: number
    totalApplicationCDDReviewAuthorise: number
    totalApplicationVerificator: number
}